board 영역 jsp css / js 에 client 폴더 경로 생성 css 분리 사이트매쉬 적용(topMenu - fixed로 변경)

+++ secondProject/src/kr/co/tripadvisor/repository/domain/BoardImage.java
@@ -6,7 +6,7 @@ public class BoardImage {
     private String path;
     private String oriName;
     private String sysName;
-    private Integer fileSize;
+    private long fileSize;
 
     public Integer getImageNo() {
         return imageNo;
@@ -38,10 +38,11 @@ public class BoardImage {
     public void setSysName(String sysName) {
         this.sysName = sysName;
     }
-    public Integer getFileSize() {
-        return fileSize;
-    }
-    public void setFileSize(Integer fileSize) {
-        this.fileSize = fileSize;
-    }
+	public long getFileSize() {
+		return fileSize;
+	}
+	public void setFileSize(long fileSize) {
+		this.fileSize = fileSize;
+	}
+
 }
